requirements.txt
tqdm
langchain
javascript
setuptools
openai
chardet
cchardet
chromadb==0.3.29
tiktoken
requests
setuptools
gymnasium
psutil
minecraft_launcher_lib
python-dotenv
langchain_google_genai
langchain_mistralai
langchain_anthropic
anthropic
langchain_cohere
langchain_google_vertexai
